Ensure we set the experiment version is null if assignment returns undefined. This fixes the endDate scenario where undefined was being returned and set in the cookie. Also moves most experiment methods out into a utility file to make testing in isolation possible. Test to be added later.
